apps/Adha-ai-service/requirements.txt
# Django & REST Framework
django>=4.2,<4.3
djangorestframework>=3.14.0
drf-yasg>=1.21.7
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.1
python-dateutil>=2.8.2

# Document processing
pymupdf>=1.23.0  # For PDF processing
opencv-python>=4.8.0  # For image processing
pytesseract>=0.3.10  # For OCR
pillow>=10.0.0  # For image handling

# AI & Machine Learning
openai>=1.0.0
tiktoken>=0.5.0
sentence-transformers>=2.2.2
chromadb>=0.4.22
langchain>=0.1.0

# Utilities
psutil>=5.9.0
kafka-python

# Monitoring
prometheus_client>=0.17.0
